Move retention types and interfaces to their own namespace
